Practice problem 3

+++ Python/01-Basics/tempCodeRunnerFile.py
@@ -1,6 +1,6 @@
 #write a program ti input sides of a square and print its area.
 
 
-side = float(input("Enter the sides of the square:"))
+side = int(input("Enter the sides of the square:"))
 
 print("Area:", side*side)
